requirements.txt
yt-dlp>=2024.0.0
pydantic>=2.0.0
requests>=2.25.0
mutagen>=1.45.0
pillow>=10.0.0
